Resolved bug on windows
Windows platform builds of raku post mid-late 2024 are exhibiting a regression  in the behaviour of all in comparisons. This patch replaces code in a type constraint check that fails under this regression.
Thic change also adds a new test file to cover API related behaviors of the module. The parameter constraint for encode-headers has been added to the API test to ensure there are no unintentional changes to this in the future.